+++ AttackBehavior.java
@@ -1,6 +1,3 @@
-import java.util.Random;
-
 public interface AttackBehavior {
     void attack(DungeonCharacter enemy);
-//    public int rollD20();
 }
