+++ Dagger.java
@@ -1,6 +1,4 @@
-import java.util.Random;
-
-public class DaggerAttack implements AttackBehavior {
+public class Dagger implements AttackBehavior {
 
 
     @Override
@@ -21,18 +19,4 @@ public class DaggerAttack implements AttackBehavior {
             System.out.printf("You stab the %s with your off hand dagger for %d damage!\n", enemy.getName(), dam);
         }
     }
-
-//    @Override
-//    public int rollD20() {
-//        Random random = new Random();
-//        return random.nextInt(20) + 1;
-//    }
-//
-//
-//    private int rollD4() {
-//        Random random = new Random();
-//        return random.nextInt(4) + 1;
-//    }
-
-
 }
