+++ Quarterstaff.java
@@ -1,6 +1,4 @@
-import java.util.Random;
-
-public class MaceAttack implements AttackBehavior {
+public class Quarterstaff implements AttackBehavior {
 
     @Override
     public void attack(DungeonCharacter enemy) {
@@ -8,7 +6,7 @@ public class MaceAttack implements AttackBehavior {
         if (doesHit) {
             int dam = Dice.d6();
             enemy.subtractHitPoints(dam);
-            System.out.printf("You smash the %s for %d damage with your mace!\n", enemy.getName(), dam);
+            System.out.printf("You smash the %s for %d damage with your quarterstaff!\n", enemy.getName(), dam);
         } else {
             System.out.printf("You miss the %s with your attack.\n", enemy.getName());
         }
